+++ pcweb/components/docpage/navbar/buttons/color.py
