+++ pcweb/styles/__init__.py
@@ -1,3 +1,1 @@
-from .colors import *
-from .fonts import *
 from .styles import *
