--- pcweb/styles/colors.py
@@ -1,94 +1,0 @@
-colors = {
-    "white": "#FFFFFF",
-    "red": {
-        50: "#FFF1F1",
-        100: "#FFDADA",
-        200: "#FFADAD",
-        300: "#FF7070",
-        400: "#FF3838",
-        500: "#FF0000",
-        600: "#DB0000",
-        700: "#8B0000",
-        800: "#420000",
-        900: "#140000",
-    },
-    "orange": {
-        50: "#FFF8F1",
-        100: "#FFECDA",
-        200: "#FFD1B1",
-        300: "#FFAB7D",
-        400: "#FF7C3F",
-        500: "#FF4F00",
-        600: "#DB3D00",
-        700: "#8B1F00",
-        800: "#4C0A00",
-        900: "#170000",
-    },
-    "green": {
-        50: "#F3FFFA",
-        100: "#CEFFEE",
-        200: "#9BFFDD",
-        300: "#57FFC5",
-        400: "#00FFA8",
-        500: "#00D98F",
-        600: "#00A86E",
-        700: "#00734C",
-        800: "#00402A",
-        900: "#00170F",
-    },
-    "blue": {
-        50: "#F3F7FE",
-        100: "#DEEAFD",
-        200: "#BED5FA",
-        300: "#97BBF5",
-        400: "#6B9EED",
-        500: "#2678EB",
-        600: "#2C61B5",
-        700: "#1A4589",
-        800: "#0C2755",
-        900: "#030F23",
-    },
-    "violet": {
-        50: "#F5EFFE",
-        100: "#D5C2FB",
-        200: "#AD9BF8",
-        300: "#9580F7",
-        400: "#7E69E0",
-        500: "#5646ED",
-        600: "#2B199C",
-        700: "#20117E",
-        800: "#1A1060",
-        900: "#03030B",
-    },
-    "indigo": {
-        50: "#FAF8FB",
-        100: "#EAE4FD",
-        200: "#DACEEE",
-        300: "#AA9EC3",
-        400: "#82799E",
-        500: "#696287",
-        600: "#494369",
-        700: "#342E5C",
-        800: "#1F1944",
-        900: "#110F1F",
-    },
-    "gray": {
-        50: "#FEFDFF",
-        100: "#F4F3F6",
-        200: "#CDCCD1",
-        300: "#B3B2B9",
-        400: "#97959F",
-        500: "#777583",
-        600: "#5C5B66",
-        700: "#3E3E44",
-        800: "#242528",
-        900: "31A1A1D",
-    },
-}
-
-text_colors = {
-    "docs": {
-        "header": colors["indigo"][800],
-        "body": colors["indigo"][700],
-    }
-}
